starting with a fresh, new controller
