+++ src/main/java/ucf/assignments/ToDoListApplication.java
@@ -32,7 +32,6 @@ public class ToDoListApplication extends Application {
             primaryStage.setResizable(false);
             primaryStage.setTitle("To Do List");
             primaryStage.show();
-            ListManagerApplicationController.setup();
         } catch (IOException e) {
             e.printStackTrace();
         }
