+++ src/main/resources/ucf/assignments/ListManagerApplication.fxml
@@ -17,7 +17,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ucf.assignments.ListManagerApplicationController">
+<AnchorPane prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ucf.assignments.ListManagerApplicationController2">
     <VBox prefHeight="600.0" prefWidth="800.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         <MenuBar>
             <Menu mnemonicParsing="false" text="File">
@@ -41,34 +41,34 @@
                     </ToolBar>
                 </VBox>
                 <VBox maxHeight="1.7976931348623157E308" prefWidth="100.0">
-               <HBox prefHeight="54.0" prefWidth="605.0">
-                  <children>
-                           <Label fx:id="ListTitle" alignment="CENTER" contentDisplay="CENTER" prefHeight="33.0" prefWidth="529.0" textAlignment="CENTER">
-                               <font>
-                                   <Font size="18.0" />
-                               </font>
-                        <padding>
-                           <Insets left="80.0" />
-                        </padding>
-                        <HBox.margin>
-                           <Insets top="5.0" />
-                        </HBox.margin>
-                           </Label>
-                           <MenuButton mnemonicParsing="false" text="Edit...">
-                               <items>
-                                   <MenuItem fx:id="NewItem" mnemonicParsing="false" onAction="#NewItemClicked" text="New Item" />
-                                   <CheckMenuItem fx:id="SortAlphabetical" mnemonicParsing="false" onAction="#SortAlphaClicked" text="Sort by Alphabetical" />
-                                   <CheckMenuItem fx:id="SortDueDate" mnemonicParsing="false" onAction="#SortDueDateClicked" text="Sort by Due Date" />
-                           <CheckMenuItem fx:id="ShowComplete" mnemonicParsing="false" onAction="#ShowCompleteClicked" text="Show Complete" />
-                           <CheckMenuItem fx:id="ShowIncomplete" mnemonicParsing="false" onAction="#ShowIncompleteClicked" text="Show Incomplete" />
-                           <MenuItem fx:id="EditList" mnemonicParsing="false" onAction="#EditListClicked" text="Edit List" />
-                               </items>
-                        <HBox.margin>
-                           <Insets top="6.0" />
-                        </HBox.margin>
-                           </MenuButton>
-                  </children>
-               </HBox>
+                    <HBox prefHeight="54.0" prefWidth="605.0">
+                        <children>
+                            <Label fx:id="ListTitle" alignment="CENTER" contentDisplay="CENTER" prefHeight="33.0" prefWidth="529.0" textAlignment="CENTER">
+                                <font>
+                                    <Font size="18.0" />
+                                </font>
+                                <padding>
+                                    <Insets left="80.0" />
+                                </padding>
+                                <HBox.margin>
+                                    <Insets top="5.0" />
+                                </HBox.margin>
+                            </Label>
+                            <MenuButton mnemonicParsing="false" text="Edit...">
+                                <items>
+                                    <MenuItem fx:id="NewItem" mnemonicParsing="false" onAction="#NewItemClicked" text="New Item" />
+                                    <CheckMenuItem fx:id="SortAlphabetical" mnemonicParsing="false" onAction="#SortAlphaClicked" text="Sort by Alphabetical" />
+                                    <CheckMenuItem fx:id="SortDueDate" mnemonicParsing="false" onAction="#SortDueDateClicked" text="Sort by Due Date" />
+                                    <CheckMenuItem fx:id="ShowComplete" mnemonicParsing="false" onAction="#ShowCompleteClicked" text="Show Complete" />
+                                    <CheckMenuItem fx:id="ShowIncomplete" mnemonicParsing="false" onAction="#ShowIncompleteClicked" text="Show Incomplete" />
+                                    <MenuItem fx:id="EditList" mnemonicParsing="false" onAction="#EditListClicked" text="Edit List" />
+                                </items>
+                                <HBox.margin>
+                                    <Insets top="6.0" />
+                                </HBox.margin>
+                            </MenuButton>
+                        </children>
+                    </HBox>
                     <ListView fx:id="ItemViewer" maxHeight="1.7976931348623157E308" prefHeight="553.0" prefWidth="605.0">
                         <contextMenu>
                             <ContextMenu>
