+++ src/main/resources/ucf/assignments/SaveScene.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.text.Font?>
 
 
-<AnchorPane prefHeight="136.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ucf.assignments.ListManagerApplicationController">
+<AnchorPane prefHeight="136.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ucf.assignments.ListManagerApplicationController2">
    <children>
       <GridPane layoutX="131.0" layoutY="73.0" prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
         <columnConstraints>
@@ -24,13 +24,13 @@
           <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
         </rowConstraints>
          <children>
-            <TextField fx:id="DirectoryTextBox" promptText="Copy and paste a comupter path..." GridPane.columnIndex="1" GridPane.rowIndex="1">
+            <TextField fx:id="DirectoryTextBox" GridPane.columnIndex="1" GridPane.rowIndex="1">
                <GridPane.margin>
                   <Insets left="10.0" right="20.0" />
                </GridPane.margin>
             </TextField>
             <Label text="Path:" GridPane.halignment="RIGHT" GridPane.rowIndex="1" />
-            <Label text="Please enter a path." GridPane.columnIndex="1">
+            <Label text="Please copy and paste the path that you want to save your list to." GridPane.columnIndex="1">
                <font>
                   <Font name="System Bold" size="12.0" />
                </font>
@@ -38,7 +38,7 @@
                   <Insets left="20.0" top="20.0" />
                </GridPane.margin>
             </Label>
-            <Button fx:id="PathOK" mnemonicParsing="false" onAction="#PathOKClicked" text="OK" textAlignment="CENTER" GridPane.columnIndex="1" GridPane.halignment="RIGHT" GridPane.rowIndex="2">
+            <Button fx:id="SaveOK" mnemonicParsing="false" onAction="#SaveOKClicked" text="OK" textAlignment="CENTER" GridPane.columnIndex="1" GridPane.halignment="RIGHT" GridPane.rowIndex="2">
                <GridPane.margin>
                   <Insets right="25.0" />
                </GridPane.margin>
